requirements.txt
# Dependencias del Medical RAG
# Valentin Moreno Vasquez - Ingeniero Biomedico, Especialista en IA
#
# Instalar con:
#   pip install -r requirements.txt

# --- API ---
fastapi==0.111.0
uvicorn==0.29.0
pydantic==2.7.1

# --- RAG ---
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.20
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
sentence-transformers==3.0.1
pypdf==4.2.0

# --- Utils ---
python-dotenv==1.0.1
pytest==8.2.0
